utils/requirements.txt
uuid
typer
awesome-slugify
ffmpeg
moviepy